+++ tee-worker/omni-executor/rpc-server/src/lib.rs
@@ -6,7 +6,6 @@ mod error_code;
 mod mailer_factory;
 mod methods;
 mod middlewares;
-pub mod native_task_types;
 mod oauth2_factory;
 mod server;
 pub mod utils;
@@ -15,7 +14,6 @@ mod verify_auth;
 
 pub use auth_token_key_store::AuthTokenKeyStore;
 pub use executor_crypto::shielding_key::ShieldingKey;
-pub use native_task_types::{GasEstimateResponse, TokenCostEstimate};
 pub use server::start_server;
 
 // Removed unused hex imports - they may be used in other modules
