+++ tee-worker/omni-executor/rpc-server/src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod gas_estimation;
 pub mod paymaster;
 pub mod pumpx;
+pub mod types;
 pub mod user_op;
